guestbook/requirements.txt
django==1.8
djangorestframework==3.6.2
mysqlclient
elastic-apm

